backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
alembic==1.13.1
anyio==4.12.1
bcrypt==4.0.1
certifi==2026.1.4
click==8.3.1
fastapi==0.131.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
pydantic==2.12.5
pydantic-settings>=2.0.0
pydantic_core==2.41.5
pydantic-settings==2.13.1
python-dotenv==1.2.1
shapely==2.0.3
python-multipart==0.0.22
PyYAML==6.0.3
SQLAlchemy==2.0.46
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
psycopg2-binary
passlib[bcrypt]
PyJWT==2.8.0
email-validator 
python-jose[cryptography]
pandas==2.2.1
